The admin should be able to edit the users balance from the admin app
when the user makes a deposit or withdrawal into the admins crypto wallet, the admin has to confirm the deposit before manually updating the users balance from the admin app to reflect in the dashboard of the main site of the user

The user should get an email notification about the changes the admin made in the admin portal. So if the admin updates the balance of the user from the admin app, the user should get an email notification about the change, etc.

Overall, I am satisfied with the overall design and layout of the admin app. Thank you