I think you get the point. I'll give you a 70%
I do not plan to use KYC. Firebase Auth is okay for my project.

All I need the admin portal for is for updating of the users balance. The user has a dash board that shows:

> Deposited amount
> Total Assets
> Total withdraws made
> Wallet balance

That is basically all. All the verification of user can be done by the firebase auth, it's not really an admin portal priority but updating balance and figures is a top priority for the admin portal

The admin portal can also be used to block and unblock users

So all I expect to see in the admin app is the users profile (users information page) and the section to edit the users balance where upon updating the balance, the user gets an email notification about the action that just occurred

I'm sure this is a simple logic flow now and it can make the database schema simpler and the admin app easier to develop and debug

Any questions??