with this, you can remove all other sections of the app beside the sections that are important. I don't think we need the instrument, transactions and sentiment analyzer sections
In the wallet section of the user profile section, can you add a bold section for the current balance(s) of the user, that is the balance that is currently showing in the users FPX dashboard.

Then put an edit button beside the balance for updating the balance. If the admin updated the balance then the new balance should reflect in the FPX dashboard and the other balances should be adjusted automatically

Example:
If the user deposits $1000 into thee crypto wallet of the admin, the admin will confirm the payment and then update the deposited balance of the user from the admin app, when the balance is updated, the total assets balance will also be updated automatically and an email notification will be sent to the users email

Understood?? Now do the needful