Okay, great!
Now let me make sure that you understand the wallet section properly

Only one wallet balance should be seen. If the admin wants to update the balance of the user, the admin will have to click the edit balance button and then insert the wallet the user deposited into e.g. btc, eth, solana, etc and the amount with the admin notes

There is no btc wallet and usd wallet. It is ony one wallet per user and it is a crypto

Each successful edit should be logged and saved in the notification section of the users FPX dashboard together with the email sent to the user after each successful update

So if the user makes a deposit, the admin will update the balance of the user from the admin portal with the information of the transaction

You might want to make the UI of the wallet section look good/better. You can do a better wallet UI design I am sure of that

Now do the needful