When I clicked the adjust balance in the wallet section, I see this
It is confusing. All I want to see is my current balance, a field bar to select the asset the user deposited and a text field bar to enter the amount the user just deposited. After I type my admin notes and click apply adjustments, I should be prompted to enter my 4 digit admin pin to verify the change

The app should use current market exchange rate to convert the balance to USDT so the user can see his/her assets in one currency

So please make this easier and simpler for everyone, thank you